pyproject.toml
[project]
name = "greenledger"
version = "0.1.0"
description = "Carbon-aware AI inference CLI"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.1",
    "firebase-admin>=6.5.0",
    "supabase>=2.4.0",
    "python-dotenv>=1.0.1",
    "pydantic-settings>=2.2.1",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
    "anthropic>=0.42.0",
    "openai>=1.60.0",
    "google-genai>=1.0.0",
    "cryptography>=42.0.0",
    "rich>=13.0.0",
    "prompt-toolkit>=3.0.0",
    "click>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["cli", "apps"]